# docs/EAP/a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW ProjeX Web API'
  description: 'Web Resources Specification (A7) for ProjeX'
servers:
  - url: http://localhost:8000
    description: Production server
...
paths:
...
  "/projects":
    get:
      operationId: R301
      summary: 'R301: Display listing of projects'
      description: 'List of user projects. Access: USR'
      tags:
        - 'M03: Project Management'
      parameters:
        - name: search
          in: query
          schema:
            type: string
            default: ''
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                type: string
        '401':
          "$ref": "#/components/responses/AuthenticationException"
    post:
      operationId: R302
      summary: 'R302: Create new project'
      description: 'Form to create new Project. Access: USR'
      tags:
        - 'M03: Project Management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                description:
                  type: string
                  nullable: true
                  maxLength: 500
                isPublic:
                  type: boolean
                finishDate:
                  type: string
                  nullable: true

              required:
                - name
                - isPublic
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Project created successfully.
                required:
                  - success
                  - message
        '401':
          "$ref": "#/components/responses/AuthenticationException"
        '422':
          "$ref": "#/components/responses/ValidationException"
        '419':
          "$ref": '#/components/responses/CsrfError'
  "/projects/search":
    get:
      operationId: R303
      summary: 'R303: Search project'
      description: 'Search project by name or description. Access: USR'
      tags:
        - 'M03: Project Management'
      parameters:
        - name: search
          in: query
          schema:
            type: string
            default: ''
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: string
                required:
                  - projects
        '401':
          "$ref": "#/components/responses/AuthenticationException"
...
  "/project/{project}/task-tables/search":
    get:
      operationId: R410
      summary: 'R410: Search for tasks.'
      description: 'Searches for tasks within a specific project. Access: MEM'
      tags:
        - 'M04: Task Management'
      parameters:
        - name: project
          in: path
          required: true
          description: ''
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
            default: ''
      responses:
        '200':
          description: 'OK.'
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      taskTables:
                        type: string
                    required:
                      - taskTables
                  - type: string
        '401':
          "$ref": "#/components/responses/AuthenticationException"
        '403':
          "$ref": "#/components/responses/AuthorizationException"
        '404':
          "$ref": "#/components/responses/ModelNotFoundException"
...
components:
  responses:
    ValidationException:
      description: 'Validation error'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Errors overview.'
              errors:
                type: object
                description: 'A detailed description of each field that failed validation.'
                additionalProperties:
                  type: array
                  items:
                    type: string
            required:
            - message
            - errors
    AuthorizationException:
      description: 'Authorization error'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
            - message
    AuthenticationException:
      description: Unauthenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
            - message
    ModelNotFoundException:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
            - message
    CsrfError:
      description: 'CSRF token missing or invalid'
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "CSRF token mismatch or missing"
              code:
                type: integer
                example: 419